Rearrange SPI clock impl
