+++ esp-hal/src/soc/esp32c2/clocks.rs
@@ -761,22 +761,3 @@ impl I2cInstance {
         });
     }
 }
-
-impl SpiInstance {
-    // SPI_FUNCTION_CLOCK
-
-    fn enable_function_clock_impl(self, _clocks: &mut ClockTree, _en: bool) {}
-
-    fn configure_function_clock_impl(
-        self,
-        _clocks: &mut ClockTree,
-        _old_config: Option<SpiFunctionClockConfig>,
-        new_config: SpiFunctionClockConfig,
-    ) {
-        let _ = self;
-        SPI2::regs().clk_gate().modify(|_, w| {
-            w.mst_clk_sel()
-                .bit(matches!(new_config, SpiFunctionClockConfig::Pll40m))
-        });
-    }
-}
