+++ esp-hal/src/soc/esp32s3/clocks.rs
@@ -967,25 +967,3 @@ impl I2cInstance {
         });
     }
 }
-
-impl SpiInstance {
-    // SPI_FUNCTION_CLOCK
-
-    fn enable_function_clock_impl(self, _clocks: &mut ClockTree, _en: bool) {}
-
-    fn configure_function_clock_impl(
-        self,
-        _clocks: &mut ClockTree,
-        _old_config: Option<SpiFunctionClockConfig>,
-        new_config: SpiFunctionClockConfig,
-    ) {
-        let regs = match self {
-            SpiInstance::Spi2 => SPI2::regs(),
-            SpiInstance::Spi3 => SPI3::regs(),
-        };
-        regs.clk_gate().modify(|_, w| {
-            w.mst_clk_sel()
-                .bit(matches!(new_config, SpiFunctionClockConfig::Apb))
-        });
-    }
-}
